Fix the issue the disappearing records. Problem with Render.com server vitality.
